Module iota_bridge::abi::eth_bridge_committee

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddressEmptyCode

Custom Error type AddressEmptyCode with signature AddressEmptyCode(address) and selector 0x9996b315

BlocklistCall

Container type for all input parameters for the blocklist function with signature blocklist(address) and selector 0xe5c7160b

BlocklistReturn

Container type for all return fields from the blocklist function with signature blocklist(address) and selector 0xe5c7160b

BlocklistUpdatedFilter

CommitteeCall

Container type for all input parameters for the committee function with signature committee() and selector 0xd864e740

CommitteeIndexCall

Container type for all input parameters for the committeeIndex function with signature committeeIndex(address) and selector 0x46b01d38

CommitteeIndexReturn

Container type for all return fields from the committeeIndex function with signature committeeIndex(address) and selector 0x46b01d38

CommitteeReturn

Container type for all return fields from the committee function with signature committee() and selector 0xd864e740

CommitteeStakeCall

Container type for all input parameters for the committeeStake function with signature committeeStake(address) and selector 0x5b1adbef

CommitteeStakeReturn

Container type for all return fields from the committeeStake function with signature committeeStake(address) and selector 0x5b1adbef

ConfigCall

Container type for all input parameters for the config function with signature config() and selector 0x79502c55

ConfigReturn

Container type for all return fields from the config function with signature config() and selector 0x79502c55

ERC1967InvalidImplementation

Custom Error type ERC1967InvalidImplementation with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3

ERC1967NonPayable

Custom Error type ERC1967NonPayable with signature ERC1967NonPayable() and selector 0xb398979f

EthBridgeCommittee

FailedInnerCall

Custom Error type FailedInnerCall with signature FailedInnerCall() and selector 0x1425ea42

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address[],uint16[],uint16) and selector 0x409ac647

InitializeConfigCall

Container type for all input parameters for the initializeConfig function with signature initializeConfig(address) and selector 0xc8f55287

InitializedFilter

InvalidInitialization

Custom Error type InvalidInitialization with signature InvalidInitialization() and selector 0xf92ee8a9

Message

Message(uint8,uint8,uint64,uint8,bytes)

NoncesCall

Container type for all input parameters for the nonces function with signature nonces(uint8) and selector 0x5035bda2

NoncesReturn

Container type for all return fields from the nonces function with signature nonces(uint8) and selector 0x5035bda2

NotInitializing

Custom Error type NotInitializing with signature NotInitializing() and selector 0xd7e6bcf8

ProxiableUUIDCall

Container type for all input parameters for the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ProxiableUUIDReturn

Container type for all return fields from the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ReentrancyGuardReentrantCall

Custom Error type ReentrancyGuardReentrantCall with signature ReentrancyGuardReentrantCall() and selector 0x3ee5aeb5

UUPSUnauthorizedCallContext

Custom Error type UUPSUnauthorizedCallContext with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba

UUPSUnsupportedProxiableUUID

Custom Error type UUPSUnsupportedProxiableUUID with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4

UpdateBlocklistWithSignaturesCall

Container type for all input parameters for the updateBlocklistWithSignatures function with signature updateBlocklistWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xf6f66e98

UpgradeInterfaceVersionCall

Container type for all input parameters for the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeInterfaceVersionReturn

Container type for all return fields from the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeToAndCallCall

Container type for all input parameters for the upgradeToAndCall function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286

UpgradeWithSignaturesCall

Container type for all input parameters for the upgradeWithSignatures function with signature upgradeWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xfa72a6d0

UpgradedFilter

VerifySignaturesCall

Container type for all input parameters for the verifySignatures function with signature verifySignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xf188949c

Enums

EthBridgeCommitteeCalls

Container type for all of the contract’s call

EthBridgeCommitteeErrors

Container type for all of the contract’s custom errors

EthBridgeCommitteeEvents

Container type for all of the contract’s events

Statics

ETHBRIDGECOMMITTEE_ABI

The parsed JSON ABI of the contract.