Module iota_bridge::abi::eth_bridge_limiter

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddressEmptyCode

Custom Error type AddressEmptyCode with signature AddressEmptyCode(address) and selector 0x9996b315

CalculateAmountInUSDCall

Container type for all input parameters for the calculateAmountInUSD function with signature calculateAmountInUSD(uint8,uint256) and selector 0xe6b04b9f

CalculateAmountInUSDReturn

Container type for all return fields from the calculateAmountInUSD function with signature calculateAmountInUSD(uint8,uint256) and selector 0xe6b04b9f

CalculateWindowAmountCall

Container type for all input parameters for the calculateWindowAmount function with signature calculateWindowAmount(uint8) and selector 0xc6b478dd

CalculateWindowAmountReturn

Container type for all return fields from the calculateWindowAmount function with signature calculateWindowAmount(uint8) and selector 0xc6b478dd

ChainHourlyTransferAmountCall

Container type for all input parameters for the chainHourlyTransferAmount function with signature chainHourlyTransferAmount(uint256) and selector 0x9056e153

ChainHourlyTransferAmountReturn

Container type for all return fields from the chainHourlyTransferAmount function with signature chainHourlyTransferAmount(uint256) and selector 0x9056e153

ChainLimitsCall

Container type for all input parameters for the chainLimits function with signature chainLimits(uint8) and selector 0x2c4535af

ChainLimitsReturn

Container type for all return fields from the chainLimits function with signature chainLimits(uint8) and selector 0x2c4535af

CommitteeCall

Container type for all input parameters for the committee function with signature committee() and selector 0xd864e740

CommitteeReturn

Container type for all return fields from the committee function with signature committee() and selector 0xd864e740

CurrentHourCall

Container type for all input parameters for the currentHour function with signature currentHour() and selector 0x020e185d

CurrentHourReturn

Container type for all return fields from the currentHour function with signature currentHour() and selector 0x020e185d

ERC1967InvalidImplementation

Custom Error type ERC1967InvalidImplementation with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3

ERC1967NonPayable

Custom Error type ERC1967NonPayable with signature ERC1967NonPayable() and selector 0xb398979f

EthBridgeLimiter

FailedInnerCall

Custom Error type FailedInnerCall with signature FailedInnerCall() and selector 0x1425ea42

GetChainHourTimestampKeyCall

Container type for all input parameters for the getChainHourTimestampKey function with signature getChainHourTimestampKey(uint8,uint32) and selector 0x085714c5

GetChainHourTimestampKeyReturn

Container type for all return fields from the getChainHourTimestampKey function with signature getChainHourTimestampKey(uint8,uint32) and selector 0x085714c5

HourlyTransferAmountUpdatedFilter

InitializeCall

Container type for all input parameters for the initialize function with signature initialize(address,uint8[],uint64[]) and selector 0x43a5f2bc

InitializedFilter

InvalidInitialization

Custom Error type InvalidInitialization with signature InvalidInitialization() and selector 0xf92ee8a9

LimitUpdatedFilter

Message

Message(uint8,uint8,uint64,uint8,bytes)

NoncesCall

Container type for all input parameters for the nonces function with signature nonces(uint8) and selector 0x5035bda2

NoncesReturn

Container type for all return fields from the nonces function with signature nonces(uint8) and selector 0x5035bda2

NotInitializing

Custom Error type NotInitializing with signature NotInitializing() and selector 0xd7e6bcf8

OldestChainTimestampCall

Container type for all input parameters for the oldestChainTimestamp function with signature oldestChainTimestamp(uint8) and selector 0x795334ad

OldestChainTimestampReturn

Container type for all return fields from the oldestChainTimestamp function with signature oldestChainTimestamp(uint8) and selector 0x795334ad

OwnableInvalidOwner

Custom Error type OwnableInvalidOwner with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7

OwnableUnauthorizedAccount

Custom Error type OwnableUnauthorizedAccount with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

OwnershipTransferredFilter

ProxiableUUIDCall

Container type for all input parameters for the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

ProxiableUUIDReturn

Container type for all return fields from the proxiableUUID function with signature proxiableUUID() and selector 0x52d1902d

RecordBridgeTransfersCall

Container type for all input parameters for the recordBridgeTransfers function with signature recordBridgeTransfers(uint8,uint8,uint256) and selector 0x9373d391

ReentrancyGuardReentrantCall

Custom Error type ReentrancyGuardReentrantCall with signature ReentrancyGuardReentrantCall() and selector 0x3ee5aeb5

RenounceOwnershipCall

Container type for all input parameters for the renounceOwnership function with signature renounceOwnership() and selector 0x715018a6

TransferOwnershipCall

Container type for all input parameters for the transferOwnership function with signature transferOwnership(address) and selector 0xf2fde38b

UUPSUnauthorizedCallContext

Custom Error type UUPSUnauthorizedCallContext with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba

UUPSUnsupportedProxiableUUID

Custom Error type UUPSUnsupportedProxiableUUID with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4

UpdateLimitWithSignaturesCall

Container type for all input parameters for the updateLimitWithSignatures function with signature updateLimitWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0x97c39b13

UpgradeInterfaceVersionCall

Container type for all input parameters for the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeInterfaceVersionReturn

Container type for all return fields from the UPGRADE_INTERFACE_VERSION function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc

UpgradeToAndCallCall

Container type for all input parameters for the upgradeToAndCall function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286

UpgradeWithSignaturesCall

Container type for all input parameters for the upgradeWithSignatures function with signature upgradeWithSignatures(bytes[],(uint8,uint8,uint64,uint8,bytes)) and selector 0xfa72a6d0

UpgradedFilter

WillAmountExceedLimitCall

Container type for all input parameters for the willAmountExceedLimit function with signature willAmountExceedLimit(uint8,uint8,uint256) and selector 0xf7857a19

WillAmountExceedLimitReturn

Container type for all return fields from the willAmountExceedLimit function with signature willAmountExceedLimit(uint8,uint8,uint256) and selector 0xf7857a19

WillUSDAmountExceedLimitCall

Container type for all input parameters for the willUSDAmountExceedLimit function with signature willUSDAmountExceedLimit(uint8,uint256) and selector 0x4f46cba7

WillUSDAmountExceedLimitReturn

Container type for all return fields from the willUSDAmountExceedLimit function with signature willUSDAmountExceedLimit(uint8,uint256) and selector 0x4f46cba7

Enums

EthBridgeLimiterCalls

Container type for all of the contract’s call

EthBridgeLimiterErrors

Container type for all of the contract’s custom errors

EthBridgeLimiterEvents

Container type for all of the contract’s events

Statics

ETHBRIDGELIMITER_ABI

The parsed JSON ABI of the contract.