iota_bridge

Module events

This file contains the definition of the IotaBridgeEvent enum, of which each variant is an emitted Event struct defined in the Move Bridge module. We rely on structures in this file to decode the bcs content of the emitted events.

Structs

BlocklistValidatorEvent

CommitteeMemberUrlUpdateEvent

CommitteeUpdate

EmergencyOpEvent

EmittedIotaToEthTokenBridgeV1

MoveBlocklistValidatorEvent

MoveCommitteeMemberUrlUpdateEvent

MoveCommitteeUpdateEvent

MoveNewTokenEvent

MoveTokenDepositedEvent

MoveTokenRegistrationEvent

MoveTokenTransferAlreadyApproved

MoveTokenTransferAlreadyClaimed

MoveTokenTransferApproved

MoveTokenTransferClaimed

MoveTokenTransferLimitExceed

MoveUpdateRouteLimitEvent

NewTokenEvent

TokenRegistrationEvent

TokenTransferAlreadyApproved

TokenTransferAlreadyClaimed

TokenTransferApproved

TokenTransferClaimed

TokenTransferLimitExceed

UpdateTokenPriceEvent

Enums

IotaBridgeEvent

Statics

BlocklistValidatorEvent

CommitteeMemberRegistration

CommitteeMemberUrlUpdateEvent

CommitteeUpdateEvent

EmergencyOpEvent

IotaToEthTokenBridgeV1

NewTokenEvent

TokenRegistrationEvent

TokenTransferAlreadyApproved

TokenTransferAlreadyClaimed

TokenTransferApproved

TokenTransferClaimed

TokenTransferLimitExceed

UpdateTokenPriceEvent