Module iota_bridge::types

Structs

AddTokensOnEvmAction

AddTokensOnIotaAction

AssetPriceUpdateAction

BlocklistCommitteeAction

BridgeActionDigest

BridgeActionKey

BridgeAuthority

BridgeCommittee

BridgeCommitteeValiditySignInfo

BridgeEventDigest

EmergencyAction

EthLog

EthToIotaBridgeAction

EthTransactionHash

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

EvmContractUpgradeAction

IotaToEthBridgeAction

LimitUpdateAction

ParsedTokenTransferMessage

RawEthLog

The version of EthLog that does not have log_index_in_tx.

Enums

BlocklistType

BridgeAction

The type of actions Bridge Committee verify and sign off to execution. Its relationship with BridgeEvent is similar to the relationship between

BridgeActionStatus

BridgeActionType

EmergencyActionType

Constants

BRIDGE_AUTHORITY_TOTAL_VOTING_POWER

BRIDGE_PAUSED

BRIDGE_UNPAUSED

USD_MULTIPLIER

Traits

EthEvent

Functions

is_route_valid

Check if the bridge route is valid Only mainnet can bridge to mainnet, other than that we do not care.

Type Aliases

CertifiedBridgeAction

IsBridgePaused

SignedBridgeAction

VerifiedCertifiedBridgeAction

VerifiedSignedBridgeAction